pyproject.toml
[tool.poetry]
name = "clv-defactio"
version = "0.1.0"
description = "Customer Lifetime Value Prediction System"
authors = ["Rob Ford <rob.ford@defacto.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11.7,<3.12"
pymc = "^5.10.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
scipy = "^1.10.0"
arviz = "^0.16.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
scikit-learn = "^1.3.0"
google-cloud-storage = "^2.10.0"
google-cloud-bigquery = "^3.11.0"
google-cloud-bigquery-storage = "^2.24.0"
kubernetes = "^28.1.0"
pyyaml = "^6.0.1"
google-auth = "^2.36.0"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
db-dtypes = "^1.3.1"
shapely = "^2.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3